Fix unconfirmed move to reject invalid move
