+++ tests/conftest.py
@@ -81,3 +81,15 @@ def fx_other_session(fx_app, fx_other_app):
     fx_db.create_all()
     fx_db.init_app(fx_app)
     return session
+
+
+@pytest.fixture
+def fx_novice_status():
+    return {
+        'strength': '13',
+        'dexterity': '12',
+        'constitution': '16',
+        'intelligence': '10',
+        'wisdom': '8',
+        'charisma': '5'
+    }
