+++ tests/game_test.py
@@ -1,5 +1,5 @@
 from nekoyume.models import Move, User
-
+from nekoyume.game import get_unconfirmed_move
 
 def test_login(fx_test_client):
     rv = fx_test_client.post('/login', data={
@@ -59,3 +59,13 @@ def test_logout(fx_test_client, fx_session, fx_user):
     rv = fx_test_client.get('/logout')
     rv = fx_test_client.get('/')
     assert rv.status == '302 FOUND'
+
+
+def test_get_unconfirmed_move(fx_session, fx_user, fx_novice_status):
+    assert not get_unconfirmed_move(fx_user.address)
+    move = fx_user.create_novice(fx_novice_status)
+    assert get_unconfirmed_move(fx_user.address)
+    # invalid move should not be recognized as unconfirmed move
+    move.id = ('00000000000000000000000000000000'
+               '00000000000000000000000000000000')
+    assert not get_unconfirmed_move(fx_user.address)
